Setup routing using react router
