+++ src/index.js
@@ -1,11 +1,21 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
+import "./index.scss";
+import App from "./App";
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+/**
+ * Starting v6 router can be used only by wrapping the whole <App />
+ * in a <BrowserRouter /> component from react-router-dom. It will
+ * allow us to use the APIs for the routing defined in the library.
+ */
 root.render(
   <React.StrictMode>
-    <App />
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
   </React.StrictMode>
 );
